# PERF: 30% faster is_period_object checks