--- a/pandas/_libs/tslibs/period.pxd
+++ b/pandas/_libs/tslibs/period.pxd
@@ -1,0 +1,1 @@
+cdef bint is_period_object(object obj)